resources: Remove Bitcoin Foundation
At present on our Resources page, the first link is to the Bitcoin
Foundation. That website has fallen into disrepair and is currently
broken.

Each day, new users are discovering Bitcoin for the first time,
along with others who are learning more about it via Bitcoin.org. This
PR removes the link to the Bitcoin Foundation, for now, so we aren't
providing a top resource pointing to a broken web page.