Wallets: fix accidental wallet hiding on category change
Closes #1493

When a wallet was selected on mobile displays, its entry was copied to
another part of the page with its element id intact.  When a category
was changed, it prevented duplicate entries from being shown by
searching the page for matching ids, which unintentionally matched the
copied element.

This patch changes the id of the copied entry during the copy operation.