Update setting-up-your-environment.md
Gave a more specific example for what the faster build can be used for. It saves much more time when you're only tinkering with html and css. I wish I had tried it sooner because the full preview build takes a lot time now, perhaps because of the new translations added