gitignore: Adds '.jekyll-cache/'
We will be updating Jekyll soon, and the newer version utilizes a
caching API. This adds its cache folder (.jekyll-cache), which is set
aside for site builds, to the gitignore file to ensure that it isn't
getting tracked along with  repository changes and accidentally included
in people's pull requests.

This will be merged once tests pass.